refactor: calculator action, type, message 분리
