+++ src/constants/message.ts
@@ -1,0 +1,7 @@
+export const MESSAGE = {
+  OVER_MAX_NUMBER_SIZE: (size: number) =>
+    `${size}자리 이상은 입력할 수 없어요.`,
+  EMPTY_NUMBER: `계산할 값을 입력해주세요!`,
+  EMPTY_OPERATION: `연산자를 입력해주세요!`,
+  LACK_NUMBER: `연산할 두 숫자를 입력해주세요!`,
+};
